+++ scio-smb/src/main/resources/core-site.xml
@@ -25,11 +25,11 @@
   <property>
     <!-- Supplies logical type support for writes -->
     <name>parquet.avro.write.data.supplier</name>
-    <value>org.apache.beam.sdk.extensions.smb.ParquetAvroFileOperations.LogicalTypeSupplier</value>
+    <value>org.apache.beam.sdk.extensions.smb.AvroLogicalTypeSupplier</value>
   </property>
   <property>
     <!-- Supplies logical type support for reads -->
     <name>parquet.avro.data.supplier</name>
-    <value>org.apache.beam.sdk.extensions.smb.ParquetAvroFileOperations.LogicalTypeSupplier</value>
+    <value>org.apache.beam.sdk.extensions.smb.AvroLogicalTypeSupplier</value>
   </property>
 </configuration>
